Add use of the default doc root as a pitfall
Using the default document root in Ubuntu, Debian, or any other operating system opens you up for data loss when the prebuilt packages update or upgrade.  As a result, as seen in bugs in Ubuntu, people are losing their code for their site by using the default document roots.  It is therefore a "pitfall" to do so, and having a blurb here to specifically state that it is a bad thing to use the default document root is logical for that reason.